+++ sound.h
@@ -5,9 +5,9 @@
 
 typedef short sample_t;
 
-void cut_off(complex double *data, complex double *Y, int low, int high, long n);
-void re_cut_off(complex double *data, complex double *Y, int low, int high, long n);
-int is_silence(complex double *data_rec, int n);
+void cut_off(complex short *data, complex double *Y, int low, int high, long n);
+void re_cut_off(complex short *data, complex double *Y, int low, int high, long n);
+int is_silence(complex short *data_rec, int n);
 void smoothing(sample_t *data, int n);
 void hamming_window(sample_t *data, int n);
 void re_hamming_window(sample_t *data, int n);
